requirements.txt
pytest-playwright==0.4.4
python-dotenv==1.0.1